Move test env. class to the package which uses it
